0006903: Made the log miner honor triggers' sync_on_insert/update/delete_conditions using similar expressions to the column match router
